examples/zero-code-examples/anthropic/requirements.txt
anthropic>=0.40.0

opentelemetry-sdk>=1.36.0
opentelemetry-exporter-otlp-proto-http>=1.36.0
opentelemetry-instrumentation-anthropic>=0.40.0
python-dotenv>=1.0.0
